# Faster binary subdivision map generation