@lukem12345 2025-02-12
Profview says that remaining bottlenecks are due to triangle gluing in the binary subdivision function. It doesn’t seem like any further performance improvements are necessary for the mesh sizes we are currently interested in, and triangle gluing is a fairly common idiom for writing high level code in this library, so we are good to merge.

The binary subdivision function could be sped up by some bookkeeping with edge indices if necessary.